Added homework lesson_11 swagger CartController + BigDecimal, Exceptions

+++ spring-web/api/src/main/java/com/geekbrains/spring/web/api/carts/CartDto.java
@@ -1,13 +1,14 @@
 package com.geekbrains.spring.web.api.carts;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class CartDto {
     private List<CartItemDto> items;
-    private int totalPrice;
+    private BigDecimal totalPrice;
 
 
-    public CartDto(List<CartItemDto> items, int totalPrice) {
+    public CartDto(List<CartItemDto> items, BigDecimal totalPrice) {
         this.items = items;
         this.totalPrice = totalPrice;
     }
@@ -19,7 +20,7 @@ public class CartDto {
         return items;
     }
 
-    public int getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
@@ -28,7 +29,7 @@ public class CartDto {
         this.items = items;
     }
 
-    public void setTotalPrice(int totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
